Move useInitDefaultRoute into a hook
