+++ packages/widget-v2/src/widget/Widget.tsx
@@ -5,7 +5,6 @@ import { createModal, useModal } from "@/components/Modal";
 import {
   cloneElement,
   ReactElement,
-  useCallback,
   useEffect,
   useMemo,
 } from "react";
@@ -13,29 +12,14 @@ import { defaultTheme, lightTheme, PartialTheme, Theme } from "./theme";
 import { Router } from "./Router";
 import { useResetAtom } from "jotai/utils";
 import { numberOfModalsOpenAtom } from "@/state/modal";
-import { useAtom, useSetAtom } from "jotai";
+import { useSetAtom } from "jotai";
 import {
-  skipAssetsAtom,
   skipClientConfigAtom,
   themeAtom,
   defaultSkipClientConfig,
 } from "@/state/skipClient";
 import { SkipClientOptions } from "@skip-go/client";
-import {
-  destinationAssetAmountAtom,
-  destinationAssetAtom,
-  sourceAssetAmountAtom,
-  sourceAssetAtom,
-} from "@/state/swapPage";
-
-export type DefaultRouteConfig = {
-  amountIn?: number;
-  amountOut?: number;
-  srcChainId?: string;
-  srcAssetDenom?: string;
-  destChainId?: string;
-  destAssetDenom?: string;
-};
+import { DefaultRouteConfig, useInitDefaultRoute } from "./useInitDefaultRoute";
 
 export type WidgetProps = {
   theme?: PartialTheme | "light" | "dark";
@@ -52,23 +36,9 @@ export const Widget = (props: WidgetProps) => {
 };
 
 const WidgetWithoutNiceModalProvider = (props: WidgetProps) => {
+  useInitDefaultRoute(props.defaultRoute);
   const setSkipClientConfig = useSetAtom(skipClientConfigAtom);
   const setTheme = useSetAtom(themeAtom);
-  const setSourceAsset = useSetAtom(sourceAssetAtom);
-  const setDestinationAsset = useSetAtom(destinationAssetAtom);
-  const setSourceAssetAmount = useSetAtom(sourceAssetAmountAtom);
-  const setDestinationAssetAmount = useSetAtom(destinationAssetAmountAtom);
-
-  const [{ data: assets }] = useAtom(skipAssetsAtom);
-
-  const getClientAsset = useCallback(
-    (denom?: string, chainId?: string) => {
-      if (!denom || !chainId) return;
-      if (!assets) return;
-      return assets.find((a) => a.denom === denom && a.chainID === chainId);
-    },
-    [assets]
-  );
 
   const mergedSkipClientConfig = useMemo(() => {
     const { theme, ...skipClientConfig } = props;
@@ -95,43 +65,11 @@ const WidgetWithoutNiceModalProvider = (props: WidgetProps) => {
   useEffect(() => {
     setSkipClientConfig(mergedSkipClientConfig);
     setTheme(mergedTheme);
-    if (props.defaultRoute && assets) {
-      const {
-        srcAssetDenom,
-        srcChainId,
-        destAssetDenom,
-        destChainId,
-        amountIn,
-        amountOut,
-      } = props.defaultRoute;
-      const sourceAsset = getClientAsset(srcAssetDenom, srcChainId);
-      const destinationAsset = getClientAsset(destAssetDenom, destChainId);
-      setDestinationAsset({
-        ...destinationAsset,
-        amount: amountOut?.toString(),
-      });
-      setSourceAsset({
-        ...sourceAsset,
-        amount: amountIn?.toString(),
-      });
-      if (amountIn) {
-        setSourceAssetAmount(amountIn?.toString());
-      } else if (amountOut) {
-        setDestinationAssetAmount(amountOut?.toString());
-      }
-    }
   }, [
-    props,
-    assets,
     setSkipClientConfig,
     mergedSkipClientConfig,
     setTheme,
-    mergedTheme,
-    getClientAsset,
-    setDestinationAsset,
-    setSourceAsset,
-    setSourceAssetAmount,
-    setDestinationAssetAmount,
+    mergedTheme
   ]);
 
   return (
